Updated format for stories
