+++ rasa/shared/core/training_data/story_writer/yaml_story_writer.py
@@ -206,9 +206,10 @@ class YAMLStoryWriter(StoryWriter):
             result[KEY_USER_INTENT] = user_utterance.intent_name
 
         if hasattr(user_utterance, "inline_comment"):
-            if user_utterance.inline_comment():
+            comment = user_utterance.inline_comment()
+            if comment:
                 result.yaml_add_eol_comment(
-                    user_utterance.inline_comment(), KEY_USER_INTENT
+                    comment, KEY_USER_INTENT
                 )
 
         if user_utterance.text and (
@@ -279,11 +280,12 @@ class YAMLStoryWriter(StoryWriter):
             result[KEY_BOT_END_TO_END_MESSAGE] = action.action_text
 
         if hasattr(action, "inline_comment"):
-            if KEY_ACTION in result:
-                result.yaml_add_eol_comment(action.inline_comment(), KEY_ACTION)
-            elif KEY_BOT_END_TO_END_MESSAGE in result:
+            comment = action.inline_comment()
+            if KEY_ACTION in result and comment:
+                result.yaml_add_eol_comment(comment, KEY_ACTION)
+            elif KEY_BOT_END_TO_END_MESSAGE in result and comment:
                 result.yaml_add_eol_comment(
-                    action.inline_comment(), KEY_BOT_END_TO_END_MESSAGE
+                    comment, KEY_BOT_END_TO_END_MESSAGE
                 )
 
         return result
