+++ tests/test_model_testing.py
@@ -14,12 +14,14 @@ from rasa.core.agent import Agent
 from rasa.shared.core.events import UserUttered
 from rasa.core.test import (
     EvaluationStore,
-    WronglyClassifiedUserUtterance,
-    WronglyPredictedAction,
+    WronglyClassifiedUserUtterance
 )
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.core.training_data.story_writer.yaml_story_writer import (
     YAMLStoryWriter,
+)
+from rasa.shared.core.events import (
+    WronglyPredictedAction
 )
 import rasa.model
 import rasa.cli.utils
